requirements.txt
#python_version >= 3.8.3
discord.py == 1.6.0
aiofiles == 0.6.0
